Improved on Naming Convention and Code quality
