+++ src/shared/baseUrl.js
@@ -1,1 +1,2 @@
-export const baseUrl = 'http://starlord.hackerearth.com/books';
+//export const baseUrl = 'http://starlord.hackerearth.com/books';
+export const baseUrl = 'http://localhost:3001/books';
